requirements.txt
# Core web framework
Flask==3.1.2

# HTTP requests for API calls and web scraping
requests==2.32.5

# HTML parsing for content extraction
beautifulsoup4==4.13.5

# Environment variable management
python-dotenv==1.1.1

# Flask dependencies (automatically installed with Flask)
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.1
blinker==1.9.0
itsdangerous==2.2.0